removed int64_t support from CommandReturnValue and added a specialized constructor for int32_t, updated relevant constructors

+++ commands/reload_command.h
@@ -21,7 +21,7 @@ public:
     template<typename EntitySystemT, typename CommandSystemT, typename AllSystemsT>
     CommandReturnValue operator()(CommandSystemT& command_system, EntitySystemT& entity_system, ResourceSystem& resource_system, InputSystem& input_system, RenderingSystem& rendering_system, AllSystemsT& all_systems, Globals& globals) const
     {
-    	globals(Globals::app_needs_reload) = CommandReturnValue{static_cast<int64_t>(1)};
+    	globals(Globals::app_needs_reload) = CommandReturnValue{1};
         std::cout << "Issuing a game reload\n";
         return CommandReturnValue{0.0};
     }
